Move test to test_LocalSearch.py and implement feedback
